+++ src/core/plugins/interactionExample/index.ts
@@ -1,4 +1,5 @@
 import * as alt from 'alt-server';
+
 import { playerFuncs } from '../../server/extensions/Player';
 import { InteractionController } from '../../server/systems/interaction';
 
@@ -6,18 +7,10 @@ alt.on('interaction:DoSomething', doSomeInteracting);
 
 function GenerateInteractions() {
     InteractionController.add({
-        event: { eventName: `interaction:DoSomething`, isServer: true },
         type: 'interaction:DoSomething',
         position: { x: 402.397308, y: -1029.67, z: 29.34688 },
-        description: 'Browse Heist',
-        blip: {
-            sprite: 112,
-            color: 7,
-            shortRange: true,
-            text: 'Do Something!',
-            pos: { x: 402.397308, y: -1029.67, z: 29.34688 },
-            scale: 1
-        }
+        description: 'Neato',
+        callback: doSomeInteracting
     });
 }