+++ src/core/server/interface/Interaction.ts
@@ -1,0 +1,9 @@
+import { Vector3 } from '../../shared/interfaces/Vector';
+
+export interface Interaction {
+    description?: string;
+    position?: Vector3;
+    identifier?: string;
+    type?: string;
+    callback?: Function;
+}