+++ documentation/changelog.md
@@ -21,6 +21,7 @@ description: List of all the changes and current versions.
 + Using SWC to transpile Typescript Faster
 + Replaced SimplyMongo with EzMongoDB
 + Rebuilt Vehicle Handler
++ Simplify Interaction(s) through Server-Side Routing Only
 ```
 
 ## 1.8.2 - April 19 - May 14, 2021